Anti-abuse system
Automatically demote admin abusers, whether they rank a user with qbot or manually.
