+++ .replit
@@ -1,0 +1,1 @@
+run = "node src/index.js"
