--- commands/setrank.js
@@ -1,151 +1,0 @@
-const roblox = require('noblox.js');
-const Discord = require('discord.js');
-const path = require('path');
-require('dotenv').config();
-
-const config = {
-    description: 'Ranks a user in the Roblox group.',
-    aliases: [],
-    usage: '<username> <role name/rank>',
-    rolesRequired: ['Ranking Permissions'],
-    category: 'Ranking'
-}
-
-let getRankFromName = async (func_rankname, func_group) => {
-    let roles = await roblox.getRoles(func_group);
-    let role = await roles.find(rank => rank.name.toLowerCase() === func_rankname.toLowerCase());
-    if(!role) {
-        return null;
-    } else {
-        return role.rank;
-    }
-}
-
-module.exports = {
-    config,
-    run: async (client, message, args) => {
-        let embed = new Discord.MessageEmbed();
-
-        let username = args[0];
-        if(!username) {
-            embed.setDescription(`Missing arguments.\n\nUsage: \`${client.config.prefix}${path.basename(__filename).split('.')[0]}${' ' + config.usage || ''}\``);
-            embed.setColor(client.config.colors.error);
-            embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-            return message.channel.send(embed);
-        }
-
-        let rank = args[1];
-        if(!rank) {
-            embed.setDescription(`Missing arguments.\n\nUsage: \`${client.config.prefix}${path.basename(__filename).split('.')[0]}${' ' + config.usage || ''}\``);
-            embed.setColor(client.config.colors.error);
-            embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-            return message.channel.send(embed);
-        }
-
-        if(/[^0-9]+/gm.test(rank)) {
-            let rankArgs = args.slice(1).join(' ');
-            if(!rankArgs) {
-                embed.setDescription(`Missing (or invalid) arguments.\n\nUsage: \`${client.config.prefix}${path.basename(__filename).split('.')[0]}${' ' + config.usage || ''}\``);
-                embed.setColor(client.config.colors.error);
-                embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-                return message.channel.send(embed);
-            }
-            let rankSearch = await getRankFromName(rankArgs, client.config.groupId);
-            if(!rankSearch) {
-                embed.setDescription('The specified rank does not exist.');
-                embed.setColor(client.config.colors.error);
-                embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-                return message.channel.send(embed);
-            }
-            rank = rankSearch;
-        }
-
-        let id;
-        try {
-            id = await roblox.getIdFromUsername(username);
-        } catch (err) {
-            embed.setDescription(`${username} is not a Roblox user.`);
-            embed.setColor(client.config.colors.error);
-            embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-            return message.channel.send(embed);
-        }
-
-        let rankInGroup = await roblox.getRankInGroup(client.config.groupId, id);
-        let rankingTo = rankInGroup - 1;
-        if(client.config.maximumRank <= rankInGroup || client.config.maximumRank <= rank) {
-            embed.setDescription('This bot cannot rank this user due to the maximum rank configured.');
-            embed.setColor(client.config.colors.error);
-            embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-            return message.channel.send(embed);
-        }
-
-        if(rankInGroup === 0){
-            embed.setDescription('That user is not in the group.');
-            embed.setColor(client.config.colors.error);
-            embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-            return message.channel.send(embed);
-        }
-
-        if(client.config.verificationChecks === true) {
-            let linkedUser = await client.utils.getLinkedUser(message.author.id, message.guild.id);
-            if(!linkedUser) {
-                embed.setDescription('You must be verified on either of the sites below to use this command.\n\n**Bloxlink:** https://blox.link\n**RoVer:** https://verify.eryn.io');
-                embed.setColor(client.config.colors.error);
-                embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-                return message.channel.send(embed);
-            }
-
-            if(linkedUser === 'RATE_LIMITS') {
-                embed.setDescription('Verification checks are currently on cooldown.');
-                embed.setColor(client.config.colors.error);
-                embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-                return message.channel.send(embed);
-            }
-
-            if(linkedUser === id) {
-                embed.setDescription('You can\'t rank yourself!');
-                embed.setColor(client.config.colors.error);
-                embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-                return message.channel.send(embed);
-            }
-
-            let linkedUserRankInGroup = await roblox.getRankInGroup(client.config.groupId, linkedUser);
-            if(rankInGroup >= linkedUserRankInGroup || rank >= linkedUserRankInGroup) {
-                embed.setDescription('You can only rank people with a rank lower than yours, to a rank that is also lower than yours.');
-                embed.setColor(client.config.colors.error);
-                embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-                return message.channel.send(embed);
-            }
-        }
-
-        let rankNameInGroup = await roblox.getRankNameInGroup(client.config.groupId, id);
-        let rankingInfo;
-        try {
-            rankingInfo = await roblox.setRank(client.config.groupId, id, Number(rank));
-        } catch (err) {
-            console.log(`Error: ${err}`);
-            embed.setDescription('Oops! An unexpected error has occured. The bot owner can check the bot logs for more information.');
-            embed.setColor(client.config.colors.error);
-            embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-            return message.channel.send(embed);
-        }
-
-        embed.setDescription(`**Success!** Ranked ${username} to ${rankingInfo.name} (${rankingInfo.rank}).`);
-        embed.setColor(client.config.colors.success);
-        embed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-        message.channel.send(embed);
-
-        if(client.config.logChannelId !== 'false') {
-            let logEmbed = new Discord.MessageEmbed();
-            let logChannel = await client.channels.fetch(client.config.logChannelId);
-            logEmbed.setDescription(`**Moderator:** <@${message.author.id}> (\`${message.author.id}\`)\n**Action:** Set Rank\n**User:** ${username} (\`${id}\`)\n**Rank Change:** ${rankNameInGroup} (${rankInGroup}) -> ${rankingInfo.name} (${rankingInfo.rank})`);
-            logEmbed.setColor(client.config.colors.info);
-            logEmbed.setAuthor(message.author.tag, message.author.displayAvatarURL());
-            logEmbed.setTimestamp();
-            logEmbed.setThumbnail(`https://www.roblox.com/Thumbs/Avatar.ashx?x=150&y=150&format=png&username=${username}`);
-            return logChannel.send(logEmbed);
-        } else {
-            return;
-        }
-    }
-}
