--- config.js
@@ -1,31 +1,0 @@
-module.exports = {
-    groupId: 0,
-    maximumRank: 0,
-    prefix: 'q!',
-    logChannelId: '',
-    shoutChannelId: '',
-    auditLogChannelId: '',
-    colors: {
-        info: '#5b57d9',
-        error: '#de554e',
-        neutral: '#f2f255',
-        success: '#43d177'
-    },
-    verificationChecks: false,
-    firedRank: 1,
-    xpRankup: {
-        enabled: false,
-        roles: [
-            // Example: {
-            //     rank: 2,
-            //     xpNeeded: 10
-            // }
-        ]
-    },
-    memberCount: {
-        enabled: false,
-        groupIconURL: '',
-        channelId: '',
-        milestones: []
-    }
-}
